Refactor recipe command into focused modules
Split the 253-line recipe.rs into three focused modules:

- recipe/mod.rs (80 lines): Main command logic and orchestration
- recipe/search.rs (21 lines): Recipe finding and matching logic
- recipe/display.rs (155 lines): Table rendering and formatting utilities

Benefits:
- 68% reduction in main logic file size for better readability
- Domain-driven separation (search vs display concerns)
- Logical coupling preserved (formatting stays with table rendering)
- Cleaner architecture with focused responsibilities
- All functionality and tests preserved
