--- src/display/mod.rs
@@ -1,3 +1,0 @@
-pub mod nutrition;
-
-pub use nutrition::render_nutrition_table;
